Mualab user 1 last commit

+++ app/src/main/java/com/mualab/org/user/activity/chat/model/FirebaseUser.java
@@ -6,7 +6,7 @@ import java.io.Serializable;
 
 public class FirebaseUser implements Serializable{
     public String firebaseToken,profilePic,userName,userType,authToken;
-    public long uId,isOnline,banAdmin;
+    public int uId,isOnline,banAdmin;
     public Object lastActivity;
     @Exclude
     public boolean isChecked = false;
